+++ sdk/viem/src/custom-actions/types/wallet-verification-challenge.ts
@@ -1,0 +1,17 @@
+import type { WalletVerificationType } from "@/custom-actions/types/wallet-verification.enum.js";
+
+/**
+ * Represents a wallet verification challenge.
+ */
+export interface WalletVerificationChallenge {
+  /** The unique identifier of the challenge. */
+  id: string;
+  /** The name of the challenge. */
+  name: string;
+  /** The verification ID. */
+  verificationId: string;
+  /** The type of verification required. */
+  verificationType: WalletVerificationType;
+  /** The challenge parameters specific to the verification type. */
+  challenge: Record<string, string>;
+}
